refactor: consolidate block palette integration tests
Merge palette-toggle.test.js (4 tests) and only-blocks-filter.test.js
(3 tests) into block-palette.test.js to reduce cold-start overhead.

Refs #240
